cleaned up databse and got dsg to actually be threadable
