+++ spynnaker/spynnaker.cfg
@@ -90,6 +90,7 @@ timeScaleFactor = None
 appID = 30
 tryReboot = True
 version = None
+set_up_FPGAs = True
 
 [Routing]
 # ------
@@ -147,7 +148,7 @@ run = True
 mode = Production
 
 [Threading]
-dsg_threads = 5
+dsg_threads = 1
 
 [Database]
 create_database = False
